+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added select-chapter and select-chapter-main buttons visible only if the outline actually contains @chapter nodes.
 - Fixed cff command as per Leo's #4640
+- Added quick-find buttons to the NAV pane.
 
 ## 1.0.17